+++ src/elements/Order.php
@@ -56,6 +56,7 @@ use craft\helpers\UrlHelper;
 use craft\i18n\Locale;
 use craft\models\Site;
 use DateTime;
+use Money\Money;
 use Throwable;
 use yii\base\Exception;
 use yii\base\InvalidArgumentException;
@@ -1059,6 +1060,20 @@ class Order extends Element
      * ```
      */
     private $_customer;
+    
+    /**
+     * @var Money
+     * @see Order::setPaymentAmount() To set the order payment amount
+     * @see Order::getPaymentAmount() To get the order payment amount
+     * ---
+     * ```php
+     * echo $order->paymentAmount;
+     * ```
+     * ```twig
+     * {{ order.paymentAmount }}
+     * ```
+     */
+    private $_paymentAmount;
 
     /**
      * @inheritdoc
@@ -1434,6 +1449,8 @@ class Order extends Element
         $justPaid = $paidInFull && $this->datePaid == null;
         $justAuthorized = $authorizedInFull && $this->dateAuthorized == null;
 
+        $canComplete = ($this->getTotalAuthorized() + $this->getTotalPaid()) > 0;
+
         // If it is no longer paid in full, set datePaid to null
         if (!$paidInFull) {
             $this->datePaid = null;
@@ -1464,7 +1481,7 @@ class Order extends Element
         // If the order is now paid or authorized in full, lets mark it as complete if it has not already been.
         if (!$this->isCompleted) {
             $totalAuthorized = $this->getTotalAuthorized();
-            if ($totalAuthorized >= $this->getTotalPrice() || $paidInFull) {
+            if ($totalAuthorized >= $this->getTotalPrice() || $paidInFull || $canComplete) {
                 // We need to remove the payment source from the order now that it's paid
                 // This means the order needs new payment details for future payments: https://github.com/craftcms/commerce/issues/891
                 // Payment information is still stored in the transactions.
@@ -2217,6 +2234,32 @@ class Order extends Element
     {
         return $this->hasOutstandingBalance();
     }
+    
+    /**
+     * Returns the paymentAmount for this order.
+     *
+     * @return float
+     */
+    public function getPaymentAmount(): float
+    {
+        if ($this->_paymentAmount && $this->_paymentAmount >= 0 && $this->_paymentAmount <= $this->getOutstandingBalance()) {
+            return $this->_paymentAmount;
+        }
+
+        return $this->getOutstandingBalance();
+    }
+
+    /**
+     * Sets the orders payment amount in the order's currency. This amount is not persisted.
+     *
+     * @param float $amount
+     */
+    public function setPaymentAmount($amount)
+    {
+        $paymentCurrency = Plugin::getInstance()->getPaymentCurrencies()->getPaymentCurrencyByIso($this->getPaymentCurrency());
+        $amount = Currency::round($amount, $paymentCurrency);
+        $this->_paymentAmount = $amount;
+    }
 
     /**
      * What is the status of the orders payment
@@ -2233,7 +2276,7 @@ class Order extends Element
             return self::PAID_STATUS_PAID;
         }
 
-        if ($this->totalPaid > 0) {
+        if ($this->getTotalPaid() > 0) {
             return self::PAID_STATUS_PARTIAL;
         }
 
@@ -2357,6 +2400,7 @@ class Order extends Element
      * Returns the difference between the order amount and amount paid.
      *
      * @return float
+     *
      */
     public function getOutstandingBalance(): float
     {