+++ src/elements/traits/OrderElementTrait.php
@@ -609,7 +609,7 @@ trait OrderElementTrait
 
     /**
      * @param $miniTable Expects an array with rows of 'label', 'value' keys values.
-     * 
+     *
      * @return string
      */
     private function _miniTable($miniTable)