+++ CHANGELOG.md
@@ -116,6 +116,10 @@
 - Emails and PDFs now have Language settings that can be used to specify the language that should be used, instead of the order’s language. ([#1884](https://github.com/craftcms/commerce/issues/1884))
 - Added the `cp.commerce.order.content`, `cp.commerce.order.edit.order-actions`, and `cp.commerce.order.edit.order-secondary-actions` template hooks to the Edit Order page. ([#138](https://github.com/craftcms/commerce/issues/138), [#1269](https://github.com/craftcms/commerce/issues/1269))
 
+### Added
+- Partial payments on orders can now be made from the Edit Order page.
+- Added the `allowFrontEndPartialPayments` config setting.
+
 ### Changed
 - Improved the Edit Product page load time by lazy-loading variants’ related sales on scroll. ([#1883](https://github.com/craftcms/commerce/issues/1883))
 - The Edit Order page no longer requires orders to have at least one line item to be saved.